Changing wording about pivot fields
Pivot fields only show up when the relationship has been defined on both sides. The documentation makes it seem like this optional. 

I have tried to add information about this, but perhaps the sentence "Of course, it is likely we would also define this field on the inverse of the relationship" should be changed as well.